Join crates and crate_owners on crate_id instead of owner_id
I noticed that the `user-graph.csv` file only contained one entry for my name.  I could hunt it down to this line `INNER JOIN crate_owners co2 ON co2.owner_id=c2.id`, which seemed to not find any match for most of the entries. The right column seems to be `co2.crate_id`.

I fixed that for the first two COPY statements.